requirements.txt
requests
python-dotenv
numpy
pydub